oms/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
alpaca-py>=0.13.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
cryptography>=42.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
